Move out AlpsNewCodepointTest
